ref: Outsource voting to ext pt 1
